refactor: move sitemap-generator to node-app
