+++ imports/node-app/plugins/sitemap-generator/index.js
@@ -1,9 +1,9 @@
-import i18n from "./i18n";
-import mutations from "./mutations";
-import resolvers from "./resolvers";
-import schemas from "./schemas";
-import startup from "./startup";
-import updateSitemapTaskForShop from "./jobs/updateSitemapTaskForShop";
+import i18n from "./i18n/index.js";
+import mutations from "./mutations/index.js";
+import resolvers from "./resolvers/index.js";
+import schemas from "./schemas/index.js";
+import startup from "./startup.js";
+import updateSitemapTaskForShop from "./jobs/updateSitemapTaskForShop.js";
 
 /**
  * @summary Import and call this function to add this plugin to your API.
