+++ imports/node-app/plugins/sitemap-generator/middleware/handle-sitemap-routes.js
@@ -1,4 +1,4 @@
-import getSitemapXML from "../util/get-sitemap-xml";
+import getSitemapXML from "../util/get-sitemap-xml.js";
 
 /**
  * @name getSitemapRouteHandler
