+++ imports/node-app/plugins/sitemap-generator/mutations/index.js
@@ -1,0 +1,5 @@
+import generateSitemaps from "./generateSitemaps.js";
+
+export default {
+  generateSitemaps
+};
