+++ imports/node-app/plugins/sitemap-generator/resolvers/Query/index.js
@@ -1,0 +1,5 @@
+import sitemap from "./sitemap.js";
+
+export default {
+  sitemap
+};
