+++ imports/node-app/plugins/sitemap-generator/resolvers/index.js
@@ -1,0 +1,9 @@
+import Mutation from "./Mutation/index.js";
+import Query from "./Query/index.js";
+import Sitemap from "./Sitemap/index.js";
+
+export default {
+  Mutation,
+  Query,
+  Sitemap
+};
