+++ imports/node-app/plugins/sitemap-generator/schemas/settings.graphql
