+++ imports/plugins/included/sitemap-generator/register.js
@@ -6,6 +6,6 @@
  * temporary bridge.
  */
 import Reaction from "/imports/plugins/core/core/server/Reaction";
-import register from "./server/no-meteor/register";
+import register from "/imports/node-app/plugins/sitemap-generator/index.js";
 
 Reaction.whenAppInstanceReady(register);
