--- imports/plugins/included/sitemap-generator/server/no-meteor/resolvers/Mutation/index.js
@@ -1,5 +1,0 @@
-import generateSitemaps from "./generateSitemaps";
-
-export default {
-  generateSitemaps
-};
