--- imports/plugins/included/sitemap-generator/server/no-meteor/resolvers/Query/index.js
@@ -1,5 +1,0 @@
-import sitemap from "./sitemap";
-
-export default {
-  sitemap
-};
