--- imports/plugins/included/sitemap-generator/server/no-meteor/resolvers/index.js
@@ -1,9 +1,0 @@
-import Mutation from "./Mutation";
-import Query from "./Query";
-import Sitemap from "./Sitemap";
-
-export default {
-  Mutation,
-  Query,
-  Sitemap
-};
